Refactor system tests. (#8984)
* Refactor system tests.

Use parameterized tests for different source formats.

* flake8
